[webpack-integration] Use Sanity’s postcss import for extract css custom properties
