+++ packages/@sanity/webpack-integration/src/v3/index.js
@@ -2,15 +2,14 @@
 const fs = require('fs')
 const path = require('path')
 const dotenv = require('dotenv')
-const postcssImport = require('postcss-import')
-const postcssCssnext = require('postcss-cssnext')
 const PartResolverPlugin = require('@sanity/webpack-loader')
-const resolveStyleImport = require('./resolveStyleImport')
+const {getPostcssImportPlugin, getPostcssPlugins, getStyleResolver} = require('./postcss')
+const extractCssCustomProperties = require('./extractCssCustomProperties')
 
 const partLoaderPath = require.resolve('@sanity/webpack-loader/lib/partLoader')
 
 function getPartResolverPlugin(options) {
-  return new PartResolverPlugin(options)
+  return new PartResolverPlugin(options, extractCssCustomProperties)
 }
 
 function tryReadDotEnv(studioRootPath, configEnv) {
@@ -81,22 +80,6 @@ function getPartLoader(options) {
 
 function getLoaders(options) {
   return [getPartLoader(options)]
-}
-
-function getStyleResolver(options) {
-  return resolveStyleImport({from: options.basePath})
-}
-
-function getPostcssImportPlugin(options) {
-  const styleResolver = getStyleResolver(options)
-  const importer = postcssImport({resolve: styleResolver})
-  return importer
-}
-
-function getPostcssPlugins(options) {
-  const importer = getPostcssImportPlugin(options)
-  const nextOpts = options.cssnext
-  return [importer, postcssCssnext(nextOpts)]
 }
 
 function getConfig(options) {
